MCP/04_09/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
mcp @ git+https://github.com/modelcontextprotocol/python-sdk.git